+++ src/resampling/uncertain_values/resample_certainvalues.jl
@@ -1,13 +1,13 @@
 import ..SamplingConstraints: SamplingConstraint
-import ..UncertainValues: CertainValue
+import ..UncertainValues: CertainScalar
 
 resample(x::Number) = x
 
-resample(v::CertainValue) = v.value 
-resample(v::CertainValue, n::Int) = [v.value for i = 1:n]
+resample(v::CertainScalar) = v.value 
+resample(v::CertainScalar, n::Int) = [v.value for i = 1:n]
 
-resample(v::CertainValue, s::SamplingConstraint) = v.value 
-resample(v::CertainValue, s::SamplingConstraint, n::Int) = [v.value for i = 1:n]
+resample(v::CertainScalar, s::SamplingConstraint) = v.value 
+resample(v::CertainScalar, s::SamplingConstraint, n::Int) = [v.value for i = 1:n]
 
 constraints = [
     :(NoConstraint), 
@@ -23,8 +23,8 @@ constraints = [
 
 for constraint in constraints
     funcs = quote 
-        resample(x::CertainValue, constraint::$(constraint)) = x.value
-        resample(x::CertainValue, constraint::$(constraint), n::Int) = [x.value for i = 1:n]
+        resample(x::CertainScalar, constraint::$(constraint)) = x.value
+        resample(x::CertainScalar, constraint::$(constraint), n::Int) = [x.value for i = 1:n]
     end
     eval(funcs)
 end