+++ test/resampling/uncertain_datasets/sequential/test_resampling_sequential_increasing.jl
@@ -6,7 +6,7 @@ using StatsBase
     # Create some uncertain data with decreasing magnitude and zero overlap between values, 
     # so we're guaranteed that a strictly decreasing sequence through the dataset exists.
     N = 10
-    t = [ i <= N/2 ? CertainValue(float(i)) : UncertainValue(Normal, i, 1) for i = 1:N]
+    t = [ i <= N/2 ? CertainScalar(float(i)) : UncertainValue(Normal, i, 1) for i = 1:N]
     T = UncertainIndexDataset(t)
     iv = UncertainIndexValueDataset(t, t)