+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # UncertainData changelog
 
-## v.0.14
+## v0.15.0
+
+### Breaking changes
+
+- `CertainValue` is renamed to `CertainScalar`.
+- Some abstract types are no longer exported.
+
+### Features
+
+- More flexible inputs to `UncertainValue` constructor.
+
+### Documentation 
+
+- Shortened and improved documentation. 
+- Use regular Documenter.jl style, not mkdocs.
+
+## v0.14.1
+
+### Features
+
+- Implement sequential resampling with chunks.
+
+### Misc
+
+- Make some methods more generic (non-breaking).
+
+## v0.14.0
 
 ### Breaking changes